+++ pkg/analyzer/lib/dart/element/element.dart
@@ -2327,7 +2327,7 @@ abstract class _TmpSharedClassElement {
   /// declares a constructor. It is a compile-time error if a mixin is derived
   /// from a class whose superclass is not Object.
   /// </blockquote>
-  /// TODO(scheglov) Deprecate and remove it.
+  @Deprecated('Not useful for clients')
   bool get isValidMixin;
 
   /// Return a list containing all of the superclass constraints defined for