+++ pkg/analyzer/lib/src/dart/element/element.dart
@@ -815,6 +815,7 @@ class ClassElementImpl extends AbstractClassElementImpl {
     setModifier(Modifier.SIMPLY_BOUNDED, isSimplyBounded);
   }
 
+  @Deprecated('Not useful for clients')
   @override
   bool get isValidMixin {
     final supertype = this.supertype;
@@ -2956,6 +2957,7 @@ class EnumElementImpl extends AbstractClassElementImpl implements EnumElement {
     setModifier(Modifier.SIMPLY_BOUNDED, isSimplyBounded);
   }
 
+  @Deprecated('Not useful for clients')
   @override
   bool get isValidMixin => false;